去掉两个编译警告

+++ axvoice.ios/axvoice.ios/UtilFunctions.m
@@ -18,7 +18,7 @@
     NSData *fileData = [NSData dataWithContentsOfFile:fileName];
     
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(fileData.bytes, fileData.length, md5Buffer);
+    CC_MD5(fileData.bytes, (unsigned int)fileData.length, md5Buffer);
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++)
